+++ geotrek/infrastructure/templates/infrastructure/sql/post_10_views.sql
@@ -21,7 +21,6 @@ CREATE VIEW {{ schema_geotrek }}.v_infrastructures AS WITH v_infra AS
             {% endfor %}
             CONCAT (e.min_elevation, 'm') AS altitude,
             t.implantation_year,
-            t.condition_id,
             t.access_id,
             t.structure_id,
             t.type_id,
@@ -44,7 +43,7 @@ SELECT a.id,
        a.name_{{ lang }} AS "Name {{ lang }}",
        {% endfor %}
        b.label AS "Type",
-       c.label AS "Condition",
+       c.labels AS "Conditions",
        {% for lang in MODELTRANSLATION_LANGUAGES %}
        a.description_{{ lang }} AS "Description {{ lang }}",
        {% endfor %}
@@ -71,7 +70,19 @@ SELECT a.id,
        a.geom
 FROM v_infra a
 LEFT JOIN infrastructure_infrastructuretype b ON a.type_id = b.id
-LEFT JOIN infrastructure_infrastructurecondition c ON a.condition_id = c.id
+
+        LEFT JOIN ( WITH infrastructure_condition AS (
+                    SELECT a_1.infrastructurecondition_id,
+                    b_2.label AS labels,
+                    a_1.infrastructure_id
+                    FROM infrastructure_infrastructurecondition b_2
+                        JOIN infrastructure_infrastructure_conditions a_1 ON a_1.infrastructurecondition_id = b_2.id
+                )
+            SELECT array_to_string(array_agg(infrastructure_condition.labels), ', '::text, '_'::text)::character varying AS labels,
+            infrastructure_condition.infrastructure_id
+            FROM infrastructure_condition
+            GROUP BY infrastructure_condition.infrastructure_id) c ON a.topo_object_id = c.infrastructure_id
+
 LEFT JOIN infrastructure_infrastructureusagedifficultylevel d ON a.usage_difficulty_id = d.id
 LEFT JOIN infrastructure_infrastructuremaintenancedifficultylevel e ON a.maintenance_difficulty_id = e.id
 LEFT JOIN common_accessmean j ON a.access_id = j.id